add new secrets API
This commit adds a new set of server APIs, client
APIs and CLI commands for secrets.

A secret is, similar to a crypto. key, an object
stored at the KES server. However, in contrast
to a key, a the value of a secret can be modified
and fetched from the client.

A typcial use case for secrets are e.g. passwords
or API keys.

This commit adds five new server APIs:
 - `/v1/secret/create`
 - `/v1/secret/describe`
 - `/v1/secret/read`
 - `/v1/secret/delete`
 - `/v1/secret/list`

Note that there is currently no modify or edit
API endpoint. Further, a secret can only be
created if no such secret exists. Otherwise,
the KES server returns a `ErrSecretExists` error.

On the client side, this commit adds new SDK
functions (corresponding to the server APIs)
and CLI commands for managing secrets.
For example:
```
$ kes secret create my-secret "password123"

$ kes secret info my-secret

$ kes secret show my-secret

$ kes secret ls

$ kes secret rm my-secret
```
See: `kes secret --help`

***

One key use case of secrets is the ability to
use a stateful KES server as keystore for a
one or multiple stateless KES servers.

Signed-off-by: Andreas Auernhammer <hi@aead.dev>
Co-authored-by: Krishnan Parthasarathi <krisis@users.noreply.github.com>
Signed-off-by: Andreas Auernhammer <hi@aead.dev>